+++ src/PHPCR/Shell/Console/Command/Phpcr/NodePropertyShowCommand.php
@@ -58,10 +58,11 @@ HERE
 
             foreach ($properties as $property) {
                 $output->writeln(sprintf(
-                    '<path>%s/</path><localname>%s</localname>: %s',
-                    $pathHelper->getParentPath($property->getPath()),
+                    '<path>%s%s</path><localname>%s</localname>: %s',
+                    $parentPath = $pathHelper->getParentPath($property->getPath()),
+                    $parentPath != '/' ? '/' : '',
                     $pathHelper->getNodeName($property->getPath()),
-                    $resultFormatHelper->formatValue($property, true)
+                    $resultFormatHelper->formatValue($property, true, false)
                 ));
             }
         }